pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kiosk-llm"
version = "0.1.0"
description = "Fine-tuned LLM Kiosk for India - Multilingual AI Assistant"
authors = [
    {name = "Advay Bhagwat", email = "bcs_2023074@iiitm.ac.in"},
    {name = "Akshita Sure", email = "bcs_2023065@iiitm.ac.in"},
    {name = "Apoorva Yadav", email = "bcs_2023076@iiitm.ac.in"},
    {name = "Dhrushit Jain", email = "bcs_2023019@iiitm.ac.in"},
    {name = "Nilanjan Mitra", email = "bcs_2023501@iiitm.ac.in"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pyaudio>=0.2.11",
    "whisper>=1.1.10",
    "sentence-transformers>=2.2.2",
    "sqlite-vec>=0.1.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "asyncio-mqtt>=0.16.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
